WIP - start over with base grapher
